backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0
